refactor(notes): relocate daily-note helpers to shared seams
Move the pure date→path math (formatIsoDate, formatDatePattern,
dailyNotePath) from the renderer's apply-template.ts to
@repo/core/notes/daily-path — the forward direction lives in core so a
future inverse (dateFromDailyPath) sits beside it and can't drift. The
template/ui-state helpers (DAILY_FOLDER_KEY, DAILY_FORMAT_KEY, defaults,
TEMPLATES_DIR, DAILY_TEMPLATE_PATH, isTemplatePath, applyTemplate) move
to the isomorphic @repo/features/daily-notes so the host's deep-link
capture drain (next stage) computes the same daily path the renderer
opens with ⌘D. Importers updated; no re-export shim.
