+++ apps/desktop/src/renderer/settings/sections/notes-section.tsx
@@ -6,7 +6,7 @@ import {
   DAILY_FORMAT_KEY,
   DEFAULT_DAILY_FOLDER,
   DEFAULT_DAILY_FORMAT,
-} from "@renderer/lib/apply-template";
+} from "@repo/features/daily-notes";
 import { useDiskState } from "@renderer/lib/use-disk-state";
 
 const parseString = (value: unknown): string | undefined =>
