+++ packages/core/package.json
@@ -16,6 +16,7 @@
     "./sync/base-store": "./src/sync/base-store.ts",
     "./sync/engine": "./src/sync/engine.ts",
     "./sync/http-sync-port": "./src/sync/http-sync-port.ts",
+    "./notes/daily-path": "./src/notes/daily-path.ts",
     "./knowledge/doc-file": "./src/knowledge/doc-file.ts",
     "./knowledge/vault-path": "./src/knowledge/vault-path.ts",
     "./knowledge/link-extract": "./src/knowledge/link-extract.ts",
